Crate y4m

YUV4MPEG2 (.y4m) Encoder/Decoder

Structs

Decoder

YUV4MPEG2 decoder.

Encoder

YUV4MPEG2 encoder.

EncoderBuilder

Encoder builder. Allows to set y4m file parameters using builder pattern.

Frame

A single frame.

Limits

Limits on the resources Decoder is allowed to use.

Ratio

Simple ratio structure since stdlib lacks one.

VendorExtensionString

A newtype wrapper around Vec to ensure validity as a vendor extension.

Enums

Colorspace

Colorspace (color model/pixel format). Only subset of them is supported.

Error

Both encoding and decoding errors.

ParseError

Granular ParseError Definiations

Functions

decode

Create a new decoder instance. Alias for Decoder::new.

encode

Create a new encoder builder. Alias for EncoderBuilder::new.